web: add logic to insert into user_deleted and host_deleted when deleting a user

+++ db/schema.sql
@@ -776,7 +776,7 @@ create table token (
 create table user_deleted (
     userid                  integer         not null,
     public_cross_project_id varchar(254)    not null,
-    create_time             double          not null,
+    create_time             integer         not null,
     primary key (userid),
     index user_deleted_userid (userid)
 ) engine=InnoDB;
@@ -784,7 +784,7 @@ create table user_deleted (
 create table host_deleted (
     hostid                  integer         not null,
     public_cross_project_id varchar(254)    not null,
-    create_time             double          not null,
+    create_time             integer         not null,
     primary key (hostid),
     index host_deleted_hostid (hostid)
 ) engine=InnoDB;
